Use Long instead of BigInteger for field offsets in MachineModel
We only compute the values as longs anyway, no need to convert them to
BigInteger.

Also use ImmutableMap, which is a more efficient implementation of a map
with guaranteed iteration order.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@25873 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/core/counterexample/AssumptionToEdgeAllocator.java
@@ -1637,13 +1637,13 @@ public class AssumptionToEdgeAllocator {
           return;
         }
 
-        Map<CCompositeTypeMemberDeclaration, BigInteger> offsets =
+        Map<CCompositeTypeMemberDeclaration, Long> offsets =
             machineModel.getAllFieldOffsetsInBits(pCompType);
 
-        for (Map.Entry<CCompositeTypeMemberDeclaration, BigInteger> memberOffset :
-            offsets.entrySet()) {
+        for (Map.Entry<CCompositeTypeMemberDeclaration, Long> memberOffset : offsets.entrySet()) {
           CCompositeTypeMemberDeclaration memberType = memberOffset.getKey();
-          handleMemberField(memberType, address.addOffset(memberOffset.getValue()));
+          handleMemberField(
+              memberType, address.addOffset(BigInteger.valueOf(memberOffset.getValue())));
         }
       }
